requirements.txt
Pillow==8.4.0
Python==3.9
